% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass.R
\name{glass}
\alias{glass}
\title{Champagne Glass Side Profile}
\usage{
glass(x)
}
\arguments{
\item{x}{A numeric value or vector of heights at which to calculate the glass radius.}
}
\value{
A numeric value or vector of radii corresponding to the heights given by `x`.
}
\description{
This function calculates the side profile of a champagne glass.
It returns the radius at a given height `x`.
}
\examples{
glass(5)
}
