% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass_purrr.R
\name{glass_purrr}
\alias{glass_purrr}
\title{Vectorized Champagne Glass Calculation using purrr}
\usage{
glass_purrr(x)
}
\arguments{
\item{x}{A numeric value or vector of heights at which to calculate the glass radius.}
}
\value{
A numeric vector of radii corresponding to the heights given by `x`.
}
\description{
This function calculates the side profile of a champagne glass using `purrr::map_dbl` for vectorized input.
}
\examples{
glass_purrr(seq(0, 20, by = 1))
}
