% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass_sapply.R
\name{glass_sapply}
\alias{glass_sapply}
\title{Vectorized Champagne Glass Calculation using sapply}
\usage{
glass_sapply(x)
}
\arguments{
\item{x}{A numeric value or vector of heights at which to calculate the glass radius.}
}
\value{
A numeric vector of radii corresponding to the heights given by `x`.
}
\description{
This function calculates the side profile of a champagne glass using `sapply` for vectorized input.
}
\examples{
glass_sapply(seq(0, 20, by = 1))
}
